Api/requirements.txt
asgiref==3.2.10
boto3==1.16.7
botocore==1.19.7
certifi==2020.6.20
cffi==1.14.3
cryptography==3.2.1
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-storages==1.10.1
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
jmespath==0.10.0
mysqlclient==2.0.1
numpy==1.19.3
Pillow==8.0.1
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.11
uWSGI==2.0.19.1
wincertstore==0.2
